Isolate EventStore persistence tests
